requirements.txt
flask==2.3.3
pillow==10.0.0
easyocr==1.7.0
gunicorn==21.2.0

# flask==2.3.3
# pillow==10.0.0
# pytesseract==0.3.10
# gunicorn==21.2.0